Started running tests for the generator
I added logging to the main file to try and quiet tensorflow, it
doesn't seem to be working. I made the variables in defaults
private. Fixed all the imports to be absolute imports. There's some
debugging code in load_image that will be removed shortly. Added the
functions to load training and testing datasets.