Finally got a single iteration printing an image
Setup the matplotlib font_manager logger to only print during error
because it's sooooo noisy. Completely changed how the default values
are stored and loaded, that's untested though. It should load from a
file during import. As long as that import is first then default
keyword parameters should be set correctly once they are
imported. Created a function to get the predicted image. Added a
couple of normalize functions, one for -1 to 1 and one for 0 to
255 and one for 0 to 1. Changed the splitting of the image to have the
correct type. Another function to string an array of images together
into a line of images. It's basically make_image_grid just with only
one row. Created a function to just run the image generator once and
print the images to a file. This is really just a proof that the code
runs. All other changes are just elpy formatting.