Finally got the training loop running
Fixed the logging for the root logger to print out all the debug
messages. Renamed Defaults to Configs, it just made sense. Added more
options for the config. Renamed the function to build the generator
and the discriminator so that variable names could make sense. Started
removing calls to config by the module itself so there's less
dependencies. Added the training and fit and checkpoints into a class
so that functions didn't take eight hundred parameters. Changed the
load_image function to a load_jpg_image function because I kept
getting weird warnings. Added a function to show some images through
matplotlib but I haven't gotten it to work yet. Added function to
build out the directory trees. The full_run function will run the
training loop from the beginning.