FINAL PRESENTATION SUBMISSION
Added my final presentation and references. There's a config for the
presentation pictures. Though this will probably be broken for anyone
other than me. Added the pictures used in the presentation.

To build the presentation you'll need emacs and a bunch of texlive
packages.

emacs ./presentation.org --batch -f org-beamer-export-to-pdf --kill

You can run the above emacs command to convert the org file to a PDF.