Added documentation and configuration
Added documentation in README.org on how to run the program. Added a
model option to the config for the load-model switch. Added the eval
code that uses torch_fidelity. I buried that in it's own file so that
torch wouldn't get imported at the sametime tensorflow is loaded. Have
a main.py inside the module to contain the actual user interaction
with the module. Loaded the checkpoints and fixed how stepping works
when the module is loaded. Fixed the image loading so that the real
and drawing images are always in the same order. Fixed load_model to
normalize correctly. Added a config file for the professor.